pyproject.toml
[project]
name = "oxapy-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "oxapy>=0.6.5",
    "python-dotenv>=1.1.1",
    "sqlalchemy-libsql>=0.2.0",
]
